pyproject.toml
[project]
name = "tts-prep-data"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "elevenlabs>=1.58.1",
    "ollama>=0.4.8",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "torch==2.5.1+cu124",
    "runorm>=1.1",
    "typer>=0.15.3",
    "openai>=1.88.0",
    "google-generativeai>=0.8.5",
]

[tool.uv.sources]
torch = { index = "pytorch-all"}
torchaudio = { index = "pytorch-all"}

[[tool.uv.index]]
name = "pytorch-all"
url = "https://download.pytorch.org/whl/"
explicit = true
